+++ src/login/logind.h
@@ -192,7 +192,7 @@ int manager_kill_unit(Manager *manager, const char *unit, KillWho who, int signo
 int manager_unit_is_active(Manager *manager, const char *unit);
 
 /* gperf lookup function */
-const struct ConfigPerfItem* logind_gperf_lookup(const char *key, unsigned length);
+const struct ConfigPerfItem* logind_gperf_lookup(const char *key, size_t length);
 
 int manager_watch_busname(Manager *manager, const char *name);
 void manager_drop_busname(Manager *manager, const char *name);
